analytics-api/pyproject.toml
[project]
name = "f1-goat-determiner"
version = "0.1.0"
description = "A tool to determine the greatest Formula 1 driver of all time based on statistics analyzis."
requires-python = ">=3.13"

dependencies = [
    "fastapi==0.122.0",
    "uvicorn==0.38.0",
    "scikit-learn==1.7.2",
    "pandas==2.3.3",
    "pydantic==2.12.5",
    "pydantic-settings==2.12.0",
    "psycopg2-binary==2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.1",
    "httpx==0.28.1",
    "ruff==0.14.7",
    "testcontainers[postgres]==4.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling==1.28.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
